Fold the MMIRS .msk parser into mmt_mmirs.py
The xfitmask .msk mask-design parser lived in its own module,
pypeit/spectrographs/mmirs_maskfile.py, but it is used only by MMIRS
(get_slitmask / get_maskdef_slitedges).  Move read_mmirs_maskfile and its
two sexagesimal helpers into mmt_mmirs.py (its only consumer) and delete the
separate module.  No behavior change; tests updated to the new location.
